tango-puzzle/backend/app/config.py
from typing import List
from pydantic import BaseModel
import os
from dotenv import load_dotenv

load_dotenv()


class Settings(BaseModel):
    PROJECT_NAME: str = "Tango Puzzle API"
    API_V1_STR: str = "/api/v1"
    
    # CORS settings - simplified for compatibility
    BACKEND_CORS_ORIGINS: List[str] = [
        "http://localhost:5173",
        "http://localhost:3000",
        "http://127.0.0.1:5173",
        "http://127.0.0.1:3000"
    ]


settings = Settings()